requirements.txt
# Core frameworks
fastapi==0.116.1
uvicorn==0.35.0
reflex==0.8.8

# Database & ORM
sqlmodel==0.0.24
sqlalchemy==2.0.43
alembic==1.16.4

# Authentication
authlib==1.6.3
starlette==0.47.2
python-multipart==0.0.20
itsdangerous==2.2.0

# Vector search & embeddings
faiss-cpu==1.12.0
sentence-transformers==5.1.0
numpy==1.26.4

# HTTP clients
httpx==0.28.1

# Environment management
python-dotenv==1.1.1

# Data processing
pandas==2.3.1

# Ollama integration
ollama==0.5.3

# Utilities
typing-extensions==4.14.1
pydantic==2.11.7
anyio==4.10.0
